chore(suite-common): move logic for default wallet backup to common code
