+++ packages/suite/src/views/onboarding/steps/SelectBackupType/LegacyOptions.tsx
@@ -1,10 +1,10 @@
+import { BackupType } from '@suite-common/suite-types';
 import { Tooltip } from '@trezor/components';
 
 import { Translation } from 'src/components/suite';
 
 import { DefaultTag } from './DefaultTag';
 import { OptionWithContent } from './OptionWithContent';
-import { BackupType } from '../../../../reducers/onboarding/onboardingReducer';
 
 type LegacyOptionsProps = {
     selected: BackupType;
