+++ suite-common/suite-types/src/walletBackupType.ts
@@ -1,0 +1,8 @@
+export const selectBackupTypes = [
+    'shamir-single',
+    'shamir-advanced',
+    '12-words',
+    '24-words',
+] as const;
+
+export type BackupType = (typeof selectBackupTypes)[number];
